feat: add Send Key step-by-step observation feature
- WaitAndRespondToCmd28: listens for TaHoma CMD 28, responds with CMD 29
  as controller (device type 1023/0x3FF), logs all frames TaHoma sends next
- LearnKeyFromController: send CMD 31 after CMD 29 (not wait to receive it)
- Web UI: Send Key button + modal with 30s countdown timer
- API: /api/send-key/start|stop|status endpoints
- Fix create_discovery_response device_type uint8_t→uint16_t (supports 0x3FF)

+++ io-homecontrol/protocol/iohome_commands.cpp
@@ -213,7 +213,7 @@ namespace iohome
         return true;
     }
 
-    bool create_discovery_response(IoFrame &frame, const uint8_t *own_node_id, const uint8_t *dst_node_id, uint8_t device_type)
+    bool create_discovery_response(IoFrame &frame, const uint8_t *own_node_id, const uint8_t *dst_node_id, uint16_t device_type)
     {
         init_frame(frame, true, true, false, false);
         set_destination(frame, dst_node_id);
